# packages/contracts/m1-identity.openapi.yaml
openapi: "3.1.0"
info:
  title: M1 Identity
  version: "1.0.0"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserOut:
      type: object
      properties:
        id: {type: integer}
        email: {type: string}
        name: {type: string}
        is_active: {type: boolean}
        role_id: {type: integer, nullable: true}
        department_id: {type: integer, nullable: true}
...
paths:
...
  /users:
    get:
      summary: List users (admin only, paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: size
          in: query
          schema: {type: integer, default: 50, maximum: 200}
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserOut"
        "403": {description: Forbidden}
    post:
      summary: Create user (admin.write)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: {type: string}
                password: {type: string}
                name: {type: string}
                role_id: {type: integer}
                department_id: {type: integer}
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /users/{id}:
    get:
      summary: Get user by ID (user.read or self)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "403": {description: Forbidden}
        "404": {description: Not found}
    patch:
      summary: Update user (user.write or self for limited fields)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string}
                department_id: {type: integer}
                role_id: {type: integer}
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
    delete:
      summary: Soft-delete user (user.delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        "204": {description: Deleted (is_active=false)}
        "403": {description: Forbidden}
...